# Optimize  `to_cupy` and `.values`