@Matt711 2025-04-24
> @Matt711 can you update the benchmark timings in your pr description with the branch-25.06 numbers too? So that we have a comparison.

Yes, will do