@Matt711 2025-04-30
> Are all of the benchmarks reported here part of our microbenchmark suite? If not, can we add them?

I'll follow-up this PR to add them. Just to clarify, the micro benchmarks are these ones here: https://github.com/rapidsai/cudf/tree/branch-25.06/python/cudf/benchmarks?